backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
langgraph
langchain
pydantic
websockets
python-dotenv
jupyter-client
ipykernel
python-multipart